util/stop: move the registry to a new file
This places the stopper registry and the test-only
`PrintLeakedStoppers` in a new file `testutils.go`.

This also ensures the test-only call to `debug.Stack` is
isolated in a separate file.

Release note: None
