+++ pkg/util/stop/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "stop",
-    srcs = ["stopper.go"],
+    srcs = [
+        "stopper.go",
+        "testutils.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/util/stop",
     visibility = ["//visibility:public"],
     deps = [
